+++ packages/client-node/__tests__/unit/node_result_set.test.ts
@@ -71,6 +71,7 @@ describe('[Node.js] ResultSet', () => {
       log_error: (err) => {
         console.error(err)
       },
+      response_headers: {},
     })
   }